internal/contour: move translator to internal/contour
Move the k8s <> envoy translator components from internal/envour to
internal/contour. This leaves internal/envoy for JSON types and
configuration support.
