+++ internal/contour/listener.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package envoy
+package contour
 
 import (
 	v2 "github.com/envoyproxy/go-control-plane/api"
